90% of automatic management of article->image links
lines 88-94 are the only remaining problem.

I need the new join-table records made. But I can't seem to get them (reliably). Sometimes the records DO get created... but almost randomly. But I've gone through so many code variations now I don't know what to think.

My theory in the current code is just add the associated image in this beforeSave() method and everything will be taken care of when the save happens.

My last change was to try returning the entity back up the chain. It suddenly seemed impossible without that. But still no joy.

If you can fix this f'n bit I'd be grateful. Here's the beginning of the section on saving associated data. Some of this I tried:

http://book.cakephp.org/3.0/en/orm/saving-data.html#saving-associations

Watch out, the loop ends on 99 and the entity is returned on 109 after the debugs