Add params to recent article search
Jason,

There are a couple of comments describing the additional query-build work that needs to be done on this.

I've made the caching much smarter to accomodate the new features but have left the query work for you.

The method works fine right now, it just doesn't do some of the new stuff.