Add topic filtering to the edit screen
I wanted to make some ‘Portal’ articles that would list articles on a topic. But what a pain to build and maintain.

The Topic feature on view pages was doing the work already for the ‘Recent Articles’ list though. It just needed different output.

The right column on edit pages had ‘resource’ section that could hold the content, so I moved the needed modules around and hooked it all up.

I have to say, this was the first time OOP really lived up to its potential in my code. It’s a real validation of the style of code in this app.

A main feature is modules that detect the conditions they need to run. Just call them or insert them and they will decide whether to act or not