Add Author input to article editing
This was a bit of a confusing mess.

I got a multi-select working but had to hard-patch the entity to get the Author association to save.

I couldn’t figure out how to get the input name to place the data at the right nest level for patchEntity. And when I manually adjusted TRD, it still wouldn’t patch.